Light refactor.
